Fix structure, code style and tests in autodiff
