+++ examples/autopoint/CMakeLists.txt
@@ -11,7 +11,6 @@ project( VolEsti )
 
 option(DISABLE_NLP_ORACLES "Disable non-linear oracles (used in collocation)" ON)
 option(BUILTIN_EIGEN "Use eigen from ../external" OFF)
-option(BUILTIN_AUTODIFF "Use eigen from ../external" OFF)
 option(USE_MKL "Use MKL library to build eigen" OFF)
 
 CMAKE_MINIMUM_REQUIRED(VERSION 3.17)
@@ -61,8 +60,15 @@ else()
 
 endif(DISABLE_NLP_ORACLES)
 
+option(BUILTIN_AUTODIFF "Use autodiff from ../external" OFF)
 include("../../external/cmake-files/Autodiff.cmake")
 GetAutodiff()
+if (BUILTIN_AUTODIFF)
+    include_directories (BEFORE ../../external/_deps/Autodiff)
+else ()
+    include_directories(BEFORE /usr/local/include)
+endif(BUILTIN_AUTODIFF)
+
 include("../../external/cmake-files/Eigen.cmake")
 GetEigen()
 
@@ -77,11 +83,7 @@ if (BUILTIN_EIGEN)
 else ()
     include_directories(BEFORE /usr/include/eigen3)
 endif(BUILTIN_EIGEN)
-if (BUILTIN_AUTODIFF)
-    include_directories (BEFORE ../../external/_deps/Autodiff)
-else ()
-    include_directories(BEFORE /usr/local/include)
-endif(BUILTIN_AUTODIFF)
+
 
 if (USE_MKL)
 find_library(BLAS
