+++ examples/autopoint/userDefinedFunction_manual_gradient.cpp
@@ -21,6 +21,7 @@
 #include "Eigen/Eigen"
 
 #include "ode_solvers/ode_solvers.hpp"
+#include "ode_solvers/oracle_autodiff_functors.hpp"
 #include "random.hpp"
 #include "random/uniform_int.hpp"
 #include "random/normal_distribution.hpp"
@@ -130,7 +131,7 @@ void run_main() {
     hmc.apply(rng, 3);
   }
   start = std::chrono::high_resolution_clock::now();
-  for (int i = 0; i < max_actual_draws; i++) { 
+  for (int i = 0; i < max_actual_draws; i++) {
     std::cout << hmc.x.getCoefficients().transpose() << std::endl;
     hmc.apply(rng, 3);
     samples.col(i) = hmc.x.getCoefficients();
